pyproject.toml
[project]
dependencies = [
  "pywebview>=5.0",
  "fastapi",
  "uvicorn[standard]",
  "python-multipart",
  "aiofiles",
  "cactus",
  "pypdf>=6.7.1",
]
name = "cactusxdeepmind-app"
requires-python = "~=3.12.0"
version = "0.1.0"

[tool.uv.sources]
cactus = {path = "./cactus/python"}
